More impls for InkFmt trait
A problem is starting to crop up: the syntax tree
produced by type-sitter makes it very hard to handle
inline comments (because they can occurr anywhere
and type-sitter can't easily account for that).

So maybe it really is easier to traverse the (untyped?)
tree with a cursor and handle nodes as they come?