Nice way of building computation nodes
Previously, we had to give two functions: One to specify keys
for the inputs and one to do a calculation on the inputs. Now
we have combinators to build computations with where you specify
a key and a computation step on the value of that key right away.

This _looks_ a bit nicer, but that means we can no longer have
a variable number of inputs. Another goal was to be able to
conditionaly include or remove values, which we also haven't
achieved. Maybe a new combinator?

But really, what I want is a way to "instrument" functions
so that the computations are only built from them *after*
all relevant keys have been registered. No clue how to
achieve that yet.