Logging now configured through cli arguments
I didn't want to spam the file system with
additional files that don't get cleaned up.

This way, users can decide to write logs to
a file if the want. Default operation is to
write the logs to stderr, which means they'll
get sent to the lsp client in lsp mode.