Update individual documents
A problem with the previous approach was that we always updated the
whole map of documents. This is, of course, completely counter to why
we’re even doing this!

So now we’re treating the individual parsers as Inputs for a document.
Had to work around the fact that we need a default impl, which isn’t too
nice.

Got rid of the twee `My<T>` wrapper; now just using `DocId`. This should
later on make it easier to move to a `DocId` that is `Copy` and has the
`URI` an an input.