Moved to mini_milc
To stress-test the “simpler” library. It’s honestly all a bit
convoluted, though arguably not *much* more than Salsa to begin with.

In the meantime, I have modified tree-sitter-ink to such an extent that
it fails to compile in its current state. That’s why we moved to
git-dependencies with exact commit hashes.