Re-implement goto-definition message
Based on a per-file list of names that state their own visibility.

This required updating `type-sitter` (due to a bug that caused infinite
loops), which in turn required updating `tree-sitter`.

The linking tests still fail to a large extent, so this needs to be
fixde. But surprisingly, the update didn’t break the formatter, which I
wouldn’t have expected.