`usage_at` improvemnts: Call nodes & prev byte
We now consider `Call` nodes in `usage_at`. This shouldn’t result in any
different behavior (calls are really just nested idents/qualified
names), but it lays the ground work for adding more context to usage
sites, which we’re probably going to use for completions.

More visibly: If we can’t find any interesting nodes at the cursor
position, we consider the previous byte instead. This enables us to find
usages right before the cursor (like when you’re currently typing a
word; tree-sitter considers that cursor position as being *outside* of
the node).

This is reflected in the completion test: We previously moved the cursor
position to the front of the string to be able to find anything. Now the
cursor position is at its old place again.