Introduce id module for document info
We want to centralize information about nodes so that we can
cross-reference it. (And ideally keep it copy, so it’s easier to work
with.)

The module is called `id` because it chiefly revolves around the
`NodeId` type. I’m not 100% confident we’ll pull it off, but the idea is
to have an copyiable, semantic ID that is location independent.

Let’s see if it works out.