Refactor: Simplify Visitor
Now in a more functional programming style again. It had this unholy mix
of imperative visitor and functional folder where the state was inside
the Visitor, but also kind of handed down to sub-visitors, which,
howewer, were constrained to be the same type … it was confusing.

All I really wanted was fix a bug where empty text in a label crasehd
the server in VSCodium (because its client library crashes on empty
labels, apparently).

This led to me trying to pass string references to sub-visitors, but
that ran afoul of the borrow checker (caused by my cobbled together
lifetime declarations).

Anyway, it was so confusing that I ended up gutting the Visitor, getting
rid of *most* of its machinery and doubling down on the “monoid”-style,
using the `combine` method.

- No more `voyage` methods, and indeed the whole `exit` machinery is
  gone as well.
- State is now entirely handled via arguments to the the `visit` and
  `traverse` methods. The methods are still `&mut self`, so they can
  still be used like normal visitors that mutate their internal state.
  But that should, by and large, not be necessary; I expect most
  problems should be tractable by the `DescendWith(…)` & `combine`
  primitives.