Separated out traversal & typed ink syntax from InkDocument
The goal is to have InkDocument only encapsulate editing, not querying.
Otherwise we’d just create boilerplate to hide the type-sitter
structures, only to then have to provide facilities to inspect/iterate
over those indirections.

Instead, we’ll use type-sitter as the lingua franca, which is the whole
point of that library.